Draggable tiles of a fixed image.
Added a log showing scrolling debug messages.

This is my second attempt at finding a way to add images to
the moving pieces: the first used an image in teh tile,
but that gave the tile a bounding box conformant with the
invisible part of the image, and also on Webkit
the image has its own dragging behaviour.
New version uses a pattertn element to map the image in to
the tile.